# namviet-backend-core/docs/swagger.yaml
definitions:
...
  customer_records.CreateCustomerVoucherRequest:
    properties:
      customer_id:
        type: string
      promotion_id:
        type: integer
      voucher_code:
        type: string
    required:
    - customer_id
    - promotion_id
    - voucher_code
    type: object
  customer_records.CreateVaccinationRecordRequest:
    properties:
      administered_by:
        type: string
      customer_id:
        type: string
      dose_number:
        type: integer
      next_due_date:
        type: string
      notes:
        type: string
      vaccination_date:
        type: string
      vaccine_name:
        type: string
    required:
    - customer_id
    - vaccination_date
    - vaccine_name
    type: object
  customer_records.CustomerVaccinationRecord:
    properties:
      administered_by:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      dose_number:
        type: integer
      id:
        type: string
      next_due_date:
        type: string
      notes:
        type: string
      vaccination_date:
        type: string
      vaccine_name:
        type: string
    type: object
  customer_records.CustomerVoucher:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      is_used:
        type: boolean
      promotion_id:
        type: integer
      used_at:
        type: string
      voucher_code:
        type: string
    type: object
  customer_records.UpdateCustomerVoucherRequest:
    properties:
      is_used:
        type: boolean
      used_at:
        type: string
    type: object
  customer_records.UpdateVaccinationRecordRequest:
    properties:
      administered_by:
        type: string
      dose_number:
        type: integer
      next_due_date:
        type: string
      notes:
        type: string
      vaccination_date:
        type: string
      vaccine_name:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Nam Viet ERP System
  title: Nam Viet ERP API
  version: "1.0"
paths:
...
  /customer-vouchers:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all customer vouchers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customer_records.CustomerVoucher'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Customer Vouchers
      tags:
      - CRM & Marketing
    post:
      consumes:
      - application/json
      description: Create a new customer voucher
      parameters:
      - description: Voucher Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer_records.CreateCustomerVoucherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer_records.CustomerVoucher'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Customer Voucher
      tags:
      - CRM & Marketing
  /customer-vouchers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a customer voucher
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Customer Voucher
      tags:
      - CRM & Marketing
    get:
      consumes:
      - application/json
      description: Retrieve a specific customer voucher
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer_records.CustomerVoucher'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Customer Voucher by ID
      tags:
      - CRM & Marketing
    put:
      consumes:
      - application/json
      description: Update customer voucher details
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer_records.UpdateCustomerVoucherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer_records.CustomerVoucher'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Customer Voucher
      tags:
      - CRM & Marketing
...
  /vaccination-records:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all vaccination records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customer_records.CustomerVaccinationRecord'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Vaccination Records
      tags:
      - CRM & Marketing
    post:
      consumes:
      - application/json
      description: Create a new vaccination record
      parameters:
      - description: Record Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer_records.CreateVaccinationRecordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer_records.CustomerVaccinationRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Vaccination Record
      tags:
      - CRM & Marketing
  /vaccination-records/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a vaccination record
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Vaccination Record
      tags:
      - CRM & Marketing
    get:
      consumes:
      - application/json
      description: Retrieve a specific vaccination record
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer_records.CustomerVaccinationRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Vaccination Record by ID
      tags:
      - CRM & Marketing
    put:
      consumes:
      - application/json
      description: Update vaccination record details
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer_records.UpdateVaccinationRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer_records.CustomerVaccinationRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Vaccination Record
      tags:
      - CRM & Marketing
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"